Fix remaining warnings: bitumen unboxing, serve/whereas, $package moves
- bitumen.TarFilesystem: ascribe `getAttribute(…).nn` scrutinee to `Any` so the
  `case n: Int`/`Long` boxed-unbox stays reachable (the agent's `java.lang.Integer`
  change broke `n & 0xfff`); `@unchecked` the partial `entry` match.
- obligatory JsonRpc dispatcher: `@unchecked` the `Optional[Text]` method match in
  the quoted body, clearing the exhaustivity warning at every `serve` call site.
- Move the `socketOptions` givens to `coaxial_core.scala` and `keywords` into
  `object VerifyTests` so neither file generates an encoded `…$package` object.
- Narrow `-Wconf` for the two genuinely-inherent cases: `Framable`'s intentional
  inline anonymous `Iterator`, and `whereas`'s deliberately-partial handler in
  `McpServer` (unmatched exceptions go to `.recover`).
