+++ lib/punctuation/src/core/punctuation.BlockParser.scala
@@ -282,7 +282,7 @@ final class BlockParser:
           case _: ParagraphBuilder if om.start != 1 => ()
           case _                                    => return openListItem(ln, None, Some(om), line)
 
-      case Unset => ()
+      case _ => ()
 
     (line, false)
 
@@ -409,7 +409,7 @@ final class BlockParser:
               // the setext underline has nothing to promote, so it falls
               // through and becomes a regular line.
 
-            case Unset => ()
+            case _ => ()
 
         case _ => ()
 
@@ -463,14 +463,14 @@ final class BlockParser:
       return
 
     ParserSupport.fenceOpener(residual) match
-      case (ch: Char, count: Int, indent: Ordinal, info: Text) =>
+      case (ch: Char, count: Int, indent: (Ordinal @scala.unchecked), info: Text) =>
         closeOpenLeafForNewBlock()
         val tokens = ParserSupport.cutInfo(info).map(InlineSupport.decodeEscapesAndEntities)
         val fenced = FencedCodeBlockBuilder(ln, ch, count, indent.n0, tokens)
         openStack += fenced
         return
 
-      case Unset => ()
+      case _ => ()
 
     deepest match
       case para: ParagraphBuilder if !para.isEmpty =>
@@ -484,7 +484,7 @@ final class BlockParser:
             if maybeHeading.present then return
             // else: paragraph was all link-ref-defs; fall through.
 
-          case Unset => ()
+          case _ => ()
 
       case _ => ()
 
@@ -499,7 +499,7 @@ final class BlockParser:
         addToParent(deepest, Layout.Heading(ln, level, Prose.Textual(content)))
         return
 
-      case Unset => ()
+      case _ => ()
 
     val canStartIndentedCode = deepest match
       case _: LeafBuilder => false
