+++ lib/telekinesis/src/test/telekinesis_test.scala
@@ -79,6 +79,8 @@ object Tests extends Suite(m"Telekinesis tests"):
       . assert(_ == t"name=Jack&age=12")
 
       test(m"Construct a Query by partial generic derivation"):
+        // This import seems to be required
+        import queryParameters.arbitraryQueryParameter
         Query.make(person = Person(t"Ken", 39)).show
 
       . assert(_ == t"person.name=Ken&person.age=39")
